chore(xaml): Refactor ResourceDictionary path resolution to shared class
